Add set and variable parsing.
Expand `Parser.parseTerm` to support variables (`Var`s) and sets (`SetLit`).
This required adding a new basetype, `UnknownT`, which is used to instantiate
values like variables and sets whose types are unknown at parse time. Unless a
better solution comes to mind we'll have to perform a pass on the full AST and
fill in missing types then.